backend/requirements.txt
flask
flask_sqlalchemy
flask_login
flask_cors
flask_bcrypt
boto3
langchain
requests
chromadb
amazon-textract-caller
psycopg2
